Add confirmation when you want to delete a post

+++ src/actions/deletePost.js
@@ -2,13 +2,17 @@ import API from '../api/api';
 import { toast } from 'react-toastify';
 
 const DeletePost = ({ id }) => {
-    API.delete(`post/${id}`)
+    const Confirm = window.confirm("Are you sure you want to delete your post ?");
+
+    if (Confirm == true) {
+        API.delete(`post/${id}`)
         .then(() => {
             toast.success('Post Deleted!');
             setTimeout(() => {
                 window.location.href = '/post';
             }, 1500);
         })
+    }  
 };
 
 export default DeletePost;
